requirements.txt
matplotlib
ipykernel
tqdm
PyQt5
fastapi
opencv-python-headless
uvicorn
fashion-clip
snoop
scikit-learn
transformers
ftfy
regex
scikit-image
pytorch_lightning==1.8.6
git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup
fastapi
torchshow
icecream
tensorboard